Make the PluggableMBeanServer the platform mbean server
