+++ jmx/src/main/java/org/jboss/as/jmx/PluggableMBeanServer.java
@@ -51,6 +51,7 @@ import javax.management.ReflectionException;
 import javax.management.loading.ClassLoaderRepository;
 
 /**
+ * An MBeanServer supporting {@link MBeanServerPlugin}s. At it's core is the original platform mbean server wrapped in TCCL behaviour.
  *
  * @author <a href="kabir.khan@jboss.com">Kabir Khan</a>
  */
@@ -60,15 +61,15 @@ public class PluggableMBeanServer implements MBeanServer {
 
     private final Set<MBeanServerPlugin> delegates = new CopyOnWriteArraySet<MBeanServerPlugin>();
 
-    public PluggableMBeanServer(MBeanServer rootMBeanServer) {
-        this.rootMBeanServer = rootMBeanServer;
+    PluggableMBeanServer(MBeanServer rootMBeanServer) {
+        this.rootMBeanServer = new TcclMBeanServer(rootMBeanServer);
     }
 
-    public void addDelegate(MBeanServerPlugin delegate) {
+    public void addPlugin(MBeanServerPlugin delegate) {
         delegates.add(delegate);
     }
 
-    public void removeDelegate(MBeanServerPlugin delegate) {
+    public void removePlugin(MBeanServerPlugin delegate) {
         delegates.remove(delegate);
     }
 
@@ -322,4 +323,267 @@ public class PluggableMBeanServer implements MBeanServer {
     public void unregisterMBean(ObjectName name) throws InstanceNotFoundException, MBeanRegistrationException {
         findDelegate(name).unregisterMBean(name);
     }
+
+    private class TcclMBeanServer implements MBeanServer {
+
+        private final MBeanServer delegate;
+
+        public TcclMBeanServer(MBeanServer delegate) {
+            this.delegate = delegate;
+        }
+
+        public void addNotificationListener(ObjectName name, NotificationListener listener, NotificationFilter filter, Object handback)
+                throws InstanceNotFoundException {
+            ClassLoader old = pushClassLoader(name);
+            try {
+                delegate.addNotificationListener(name, listener, filter, handback);
+            } finally {
+                resetClassLoader(old);
+            }
+
+        }
+
+        public void addNotificationListener(ObjectName name, ObjectName listener, NotificationFilter filter, Object handback)
+                throws InstanceNotFoundException {
+            ClassLoader old = pushClassLoader(name);
+            try {
+                delegate.addNotificationListener(name, listener, filter, handback);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public ObjectInstance createMBean(String className, ObjectName name, Object[] params, String[] signature) throws ReflectionException,
+                InstanceAlreadyExistsException, MBeanException, NotCompliantMBeanException {
+            return delegate.createMBean(className, name, params, signature);
+        }
+
+        public ObjectInstance createMBean(String className, ObjectName name, ObjectName loaderName, Object[] params, String[] signature)
+                throws ReflectionException, InstanceAlreadyExistsException, MBeanException, NotCompliantMBeanException,
+                InstanceNotFoundException {
+            ClassLoader old = pushClassLoaderByName(loaderName);
+            try {
+                return delegate.createMBean(className, name, loaderName, params, signature);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public ObjectInstance createMBean(String className, ObjectName name, ObjectName loaderName) throws ReflectionException,
+                InstanceAlreadyExistsException, MBeanException, NotCompliantMBeanException, InstanceNotFoundException {
+
+            ClassLoader old = pushClassLoaderByName(loaderName);
+            try {
+                return delegate.createMBean(className, name, loaderName);
+            } finally {
+                resetClassLoader(old);
+            }
+
+        }
+
+        public ObjectInstance createMBean(String className, ObjectName name) throws ReflectionException, InstanceAlreadyExistsException,
+                 MBeanException, NotCompliantMBeanException {
+            return delegate.createMBean(className, name);
+        }
+
+        public ObjectInputStream deserialize(ObjectName name, byte[] data) throws OperationsException {
+            return delegate.deserialize(name, data);
+        }
+
+        public ObjectInputStream deserialize(String className, byte[] data) throws OperationsException, ReflectionException {
+            return delegate.deserialize(className, data);
+        }
+
+        public ObjectInputStream deserialize(String className, ObjectName loaderName, byte[] data) throws OperationsException,
+                ReflectionException {
+            return delegate.deserialize(className, loaderName, data);
+        }
+
+        public Object getAttribute(ObjectName name, String attribute) throws MBeanException, AttributeNotFoundException, InstanceNotFoundException,
+                ReflectionException {
+            ClassLoader old = pushClassLoader(name);
+            try {
+                return delegate.getAttribute(name, attribute);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public AttributeList getAttributes(ObjectName name, String[] attributes) throws InstanceNotFoundException, ReflectionException {
+            ClassLoader old = pushClassLoader(name);
+            try {
+                return delegate.getAttributes(name, attributes);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public ClassLoader getClassLoader(ObjectName loaderName) throws InstanceNotFoundException {
+            return delegate.getClassLoader(loaderName);
+        }
+
+        public ClassLoader getClassLoaderFor(ObjectName mbeanName) throws InstanceNotFoundException {
+            return delegate.getClassLoaderFor(mbeanName);
+        }
+
+        public ClassLoaderRepository getClassLoaderRepository() {
+            return delegate.getClassLoaderRepository();
+        }
+
+        public String getDefaultDomain() {
+            return delegate.getDefaultDomain();
+        }
+
+        public String[] getDomains() {
+            return delegate.getDomains();
+        }
+
+        public Integer getMBeanCount() {
+            return delegate.getMBeanCount();
+        }
+
+        public MBeanInfo getMBeanInfo(ObjectName name) throws InstanceNotFoundException, IntrospectionException, ReflectionException {
+            return delegate.getMBeanInfo(name);
+        }
+
+        public ObjectInstance getObjectInstance(ObjectName name) throws InstanceNotFoundException {
+            return delegate.getObjectInstance(name);
+        }
+
+        public Object instantiate(String className, Object[] params, String[] signature) throws ReflectionException, MBeanException {
+            return delegate.instantiate(className, params, signature);
+        }
+
+        public Object instantiate(String className, ObjectName loaderName, Object[] params, String[] signature) throws ReflectionException,
+                MBeanException, InstanceNotFoundException {
+            ClassLoader old = pushClassLoaderByName(loaderName);
+            try {
+                return delegate.instantiate(className, loaderName, params, signature);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public Object instantiate(String className, ObjectName loaderName) throws ReflectionException, MBeanException, InstanceNotFoundException {
+            ClassLoader old = pushClassLoaderByName(loaderName);
+            try {
+                return delegate.instantiate(className, loaderName);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public Object instantiate(String className) throws ReflectionException, MBeanException {
+            return delegate.instantiate(className);
+        }
+
+        public Object invoke(ObjectName name, String operationName, Object[] params, String[] signature) throws InstanceNotFoundException,
+                MBeanException, ReflectionException {
+
+            ClassLoader old = pushClassLoader(name);
+            try {
+                return delegate.invoke(name, operationName, params, signature);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public boolean isInstanceOf(ObjectName name, String className) throws InstanceNotFoundException {
+            return delegate.isInstanceOf(name, className);
+        }
+
+        public boolean isRegistered(ObjectName name) {
+            return delegate.isRegistered(name);
+        }
+
+        public Set<ObjectInstance> queryMBeans(ObjectName name, QueryExp query) {
+            return delegate.queryMBeans(name, query);
+        }
+
+        public Set<ObjectName> queryNames(ObjectName name, QueryExp query) {
+            return delegate.queryNames(name, query);
+        }
+
+        public ObjectInstance registerMBean(Object object, ObjectName name) throws InstanceAlreadyExistsException, MBeanRegistrationException,
+                NotCompliantMBeanException {
+            return delegate.registerMBean(object, name);
+        }
+
+        public void removeNotificationListener(ObjectName name, NotificationListener listener, NotificationFilter filter, Object handback)
+                throws InstanceNotFoundException, ListenerNotFoundException {
+            ClassLoader old = pushClassLoader(name);
+            try {
+                delegate.removeNotificationListener(name, listener, filter, handback);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public void removeNotificationListener(ObjectName name, NotificationListener listener) throws InstanceNotFoundException,
+                ListenerNotFoundException {
+            ClassLoader old = pushClassLoader(name);
+            try {
+                delegate.removeNotificationListener(name, listener);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public void removeNotificationListener(ObjectName name, ObjectName listener, NotificationFilter filter, Object handback)
+                throws InstanceNotFoundException, ListenerNotFoundException {
+            ClassLoader old = pushClassLoader(name);
+            try {
+                delegate.removeNotificationListener(name, listener, filter, handback);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public void removeNotificationListener(ObjectName name, ObjectName listener) throws InstanceNotFoundException, ListenerNotFoundException {
+            ClassLoader old = pushClassLoader(name);
+            try {
+                delegate.removeNotificationListener(name, listener);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public void setAttribute(ObjectName name, Attribute attribute) throws InstanceNotFoundException, AttributeNotFoundException,
+                InvalidAttributeValueException, MBeanException, ReflectionException {
+            ClassLoader old = pushClassLoader(name);
+            try {
+                delegate.setAttribute(name, attribute);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public AttributeList setAttributes(ObjectName name, AttributeList attributes) throws InstanceNotFoundException, ReflectionException {
+            ClassLoader old = pushClassLoader(name);
+            try {
+                return delegate.setAttributes(name, attributes);
+            } finally {
+                resetClassLoader(old);
+            }
+        }
+
+        public void unregisterMBean(ObjectName name) throws InstanceNotFoundException, MBeanRegistrationException {
+            delegate.unregisterMBean(name);
+        }
+
+        private ClassLoader pushClassLoader(ObjectName name) throws InstanceNotFoundException {
+            ClassLoader mbeanCl = delegate.getClassLoaderFor(name);
+            return SecurityActions.setThreadContextClassLoader(mbeanCl);
+        }
+
+        private ClassLoader pushClassLoaderByName(ObjectName loaderName) throws InstanceNotFoundException {
+            ClassLoader mbeanCl = delegate.getClassLoader(loaderName);
+            return SecurityActions.setThreadContextClassLoader(mbeanCl);
+        }
+
+        private void resetClassLoader(ClassLoader cl) {
+            SecurityActions.resetThreadContextClassLoader(cl);
+        }
+    }
+
 }
