+++ jmx/src/main/java/org/jboss/as/jmx/SecurityActions.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.jmx.tcl;
+package org.jboss.as.jmx;
 
 import java.security.AccessController;
 import java.security.PrivilegedAction;
